backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.9
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0
plotly>=5.18.0
openpyxl>=3.1.0
reportlab>=4.0.0
groq>=0.9.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
langgraph>=0.0.40

# Install separately (may fail on some systems; skip if needed):
# lightgbm>=4.0.0
# catboost>=1.2.0
# shap>=0.44.0
# faiss-cpu>=1.7.4
# sentence-transformers>=2.6.0
# imbalanced-learn>=0.11.0
# kaleido>=0.2.1
